this is more accurate:

```
Processing time (w/o IO): 18.708785ms
Processing time (w/o IO): 18.180406ms
Processing time (w/o IO): 18.605787ms
Processing time (w/o IO): 17.920885000000002ms
Processing time (w/o IO): 18.325236ms
Processing time (w/o IO): 18.091424ms
Processing time (w/o IO): 18.554257ms
Processing time (w/o IO): 18.720408000000003ms
Processing time (w/o IO): 18.382645999999998ms
Processing time (w/o IO): 18.396163ms
```

before it had no decimal places