@jinyus 2023-11-03
```
Julia:

        Benchmark 1: julia --startup-file=no --project=Related -e "using Related; main()"
        Processing time (w/o IO): 23.912955ms
        Processing time (w/o IO): 23.974456ms
        Processing time (w/o IO): 23.908855ms
        Processing time (w/o IO): 23.826653999999998ms
        Processing time (w/o IO): 23.891055ms
        Processing time (w/o IO): 23.971059999999998ms
        Processing time (w/o IO): 24.033162ms
        Processing time (w/o IO): 23.854736000000003ms
        Processing time (w/o IO): 24.037439ms
        Processing time (w/o IO): 24.029739000000003ms
        Processing time (w/o IO): 24.164642ms
        Processing time (w/o IO): 24.02424ms
        Processing time (w/o IO): 24.04784ms
          Time (mean ± σ):      1.011 s ±  0.013 s    [User: 0.882 s, System: 0.233 s]
          Range (min … max):    0.992 s …  1.028 s    10 runs
         
Julia:

        Benchmark 1: julia --startup-file=no --project=Related -e "using Related; main()"
        Processing time (w/o IO): 383.530426ms
        Processing time (w/o IO): 383.293223ms
        Processing time (w/o IO): 382.94171700000004ms
          Time (mean ± σ):      1.463 s ±  0.027 s    [User: 1.324 s, System: 0.242 s]
          Range (min … max):    1.443 s …  1.482 s    2 runs
         
Julia:

        Benchmark 1: julia --startup-file=no --project=Related -e "using Related; main()"
        Processing time (w/o IO): 3161.8328189999997ms
        Processing time (w/o IO): 3165.394359ms
        Processing time (w/o IO): 3169.175456ms
          Time (mean ± σ):      4.629 s ±  0.002 s    [User: 4.404 s, System: 0.320 s]
          Range (min … max):    4.628 s …  4.630 s    2 runs
```